Add files via upload
Help figure out how the E[] should be written instead to avoid error. also this is just my interpretation for the resizeable array structure of a node and stack and there might be other approaches or i might have misunderstood. also i didnt do any cases for where the array is empty